ANDROID: scsi: ufs: allow overriding the blk_keyslot_manager
Add UFSHCD_QUIRK_CUSTOM_KEYSLOT_MANAGER which tells ufshcd-core that the
host controller supports inline encryption, but it needs to initialize
the crypto capabilities in a nonstandard way and/or it needs to override
blk_ksm_ll_ops.  If enabled, the standard code won't initialize the
blk_keyslot_manager; ufs_hba_variant_ops::init() must do it instead.

It is permitted that the blk_keyslot_manager has no keyslots, i.e. is a
"passthrough keyslot manager".

This is needed for FMP support, as well as for wrapped key support on
some hardware.

Bug: 166139333
Bug: 162257402
Change-Id: I757c84d3434b935faa8ba7fcf5318af91e694d04
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>