ANDROID: GKI: enable networking; add hidden network configs
Drop the explicit disable of CONFIG_ETHERNET from gki_defconfig so
wired networking support is enabled by default.

Also, when testing GKI with out-of-tree wired networking modules, it
was discovered that a few hidden Kconfigs are required.  Add them here
so out of tree net drivers can be built in a separate tree.

Bug: 179406580
Change-Id: Ia66742e514ae1a3c9f8d0ec1d92eb40b68d12ce7
Signed-off-by: Kevin Hilman <khilman@baylibre.com>
Signed-off-by: Matthias Maennich <maennich@google.com>